requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
langchain-ollama>=0.0.1
langgraph>=0.0.10
langchain-core>=0.1.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
slack-sdk>=3.26.0 